Added menu button
