+++ pages/_app.tsx
@@ -1,9 +1,10 @@
 import "@/styles/app.css";
 import type { AppProps } from "next/app";
-import { Authenticator } from '@aws-amplify/ui-react';
+import { Authenticator } from "@aws-amplify/ui-react";
 import { Amplify } from "aws-amplify";
-import outputs from "@/amplify_outputs.json";
 import "@aws-amplify/ui-react/styles.css";
+import Layout from "@/components/Layout";
+import outputs from "@/amplify_outputs.json";
 
 Amplify.configure(outputs);
 
@@ -11,7 +12,9 @@ export default function App({ Component, pageProps }: AppProps) {
   return (
     <Authenticator hideSignUp>
       {({ signOut, user }) => (
-        <Component {...pageProps} signOut={signOut} user={user} />
+        <Layout>
+          <Component {...pageProps} signOut={signOut} user={user} />
+        </Layout>
       )}
     </Authenticator>
   );
