+++ pages/page2.tsx
@@ -1,0 +1,3 @@
+export default function Page2() {
+    return <h2>TODO: Page 2 content</h2>;
+  }
