+++ pages/page3.tsx
@@ -1,0 +1,3 @@
+export default function Page3() {
+    return <h2>TODO: Page 3 content</h2>;
+  }
